requirements.txt
fastapi
pydantic[email]
pydantic-settings
uvicorn[standard]
asyncpg
redis
httpx
tenacity
fastapi-limiter
python-dotenv
pytz
